EKeeda DS June 2022/Class-73-23-12-2022-Host ML Projects with Streamlit/Bank Customer Churn Prediction/requirements.txt
joblib==0.17.0
matplotlib==3.3.2
numpy==1.23.5
pandas==1.1.3
PyYAML==6.0
scikit_learn==1.2.0
streamlit==1.16.0
